requirements.txt
flask
pandas
scikit-surprise
scikit-learn